Bug Fixed
Exception [System.OverflowException: Value was either too large or too small for an Int32.] that occurred whenever a class that inherits [Person] was instantiated

+++ Data/ViewModels/ProfessorVM.cs
@@ -6,8 +6,8 @@ namespace API.Data.ViewModels
     {
         [Required] public string senha { get; set; }
         [Required] public string nome { get; set; }
-        [Required][Range(1, int.MaxValue, ErrorMessage = "Value for {0} must be between {1} and {2}.")] public Int64 cpf { get; set; }
-        [Required][Range(1, int.MaxValue, ErrorMessage = "Value for {0} must be between {1} and {2}.")] public Int64 rg { get; set; }
+        [Required][Range(1, Int64.MaxValue, ErrorMessage = "Value for {0} must be between {1} and {2}.")] public Int64 cpf { get; set; }
+        [Required][Range(1, Int64.MaxValue, ErrorMessage = "Value for {0} must be between {1} and {2}.")] public Int64 rg { get; set; }
         public Int64? telefone { get; set; }
         public string? email { get; set; }
     }
